pyproject.toml
[tool.poetry]
name = "hand-of-hsaka"
version = "0.1.0"
description = ""
authors = ["Kaung Htet Zaw <kayhzaw@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
ipython = "^7.31.0"
tweepy = "^4.4.0"
python-dotenv = "^0.19.2"
python-binance = "^1.0.15"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
